.Update moved parsing of schedule entries to config
